Update for VC-1745
